Sample object of class ir

A sample object of class ir. The data set contains ATR-MIR spectra for a set of organic reference materials along with their metadata (types of samples and a description) and accessory data (Klason lignin mass fraction and holocellulose mass fraction).

Usage

ir_sample_data

Format

A data frame with 58 rows and 7 variables:

id_measurement

An integer vector with a unique id for each spectrum.

id_sample

A character vector with a unique id for each sample.

sample_type

A character vector containing class labels for the types of reference materials.

sample_comment

A character vector containing comments to each sample.

klason_lignin

A numeric vector with the mass fractions of Klason lignin in each sample.

holocellulose

A numeric vector with the mass fractions of holocellulose in each sample.

spectra

See ir_new_ir().

Source

The data set was derived from https://www.nature.com/articles/s41467-018-06050-2 and published by Hodgkins.2018;textualir under the CC BY 4.0 license https://creativecommons.org/licenses/by/4.0/. Hodgkins.2018;textualir originally derived the data on Klason Lignin and Holocellulose content from DelaCruz.2016;textualir.

References