Converts absorbance spectra to transmittance spectra or vice versa

ir_to_transmittance converts absorbance spectra to transmittance spectra. ir_to_absorbance converts transmittance spectra to absorbance spectra. Note that neither function checks whether the input spectra are absorbance or transmittance spectra.

Usage

ir_to_transmittance(x)

ir_to_absorbance(x)

Source

Stuart.2004ir.

Arguments

x

An object of class ir.

Value

x with y values fore each spectrum as transmittance values (in case of ir_to_transmittance) or absorbance values (in case of ir_to_absorbance).

References

Examples

# convert from absorbance to transmittance
x1 <-
    ir_sample_data |>
    ir_to_transmittance()

# convert from transmittance to absorbance
x2 <-
    x1 |>
    ir::ir_to_absorbance()

vapply(
  seq_along(x2$spectra),
  FUN = function(i) all.equal(x2$spectra[[i]], ir::ir_sample_data$spectra[[i]]),
  FUN.VALUE = logical(1L)
) |>
  all()
#> [1] TRUE