Subsetting ir objects

Subsetting ir objects

Usage

# S3 method for ir
[(x, i, j, ..., exact = TRUE)

# S3 method for ir
$(x, i)

# S3 method for ir
[[(x, i, j, ..., exact = TRUE)

# S3 method for ir
$(x, i, j, ...) <- value

# S3 method for ir
[(i, j, ..., exact = TRUE) <- value

# S3 method for ir
[[(i, j, ..., exact = TRUE) <- value

Arguments

x

An object of class ir.

i, j

Row and column indices. If j is omitted, i is used as column index.

...

Ignored.

exact

Ignored, with a warning.

value

A value to store in a row, column, range or cell. Tibbles are stricter than data frames in what is accepted here.

Value

If the subsetting operation preserves a valid spectra column (see ir()), an object of class ir with accordingly subsetted rows or columns. Else a tibble::tbl_df() or vector.

Examples

# subsetting rows
ir_sample_data[1, ]
#> # A tibble: 1 × 7
#>   id_measurement id_sample sample_type sample_comment       klason_lignin
#> *          <int> <chr>     <chr>       <chr>                          [1]
#> 1              1 GN 11-389 needles     Abies Firma Momi fir         0.360
#> # … with 2 more variables: holocellulose [1], spectra <named list>
ir_sample_data[10:15, ]
#> # A tibble: 6 × 7
#>   id_measurement id_sample sample_type sample_comment              klason_lignin
#> *          <int> <chr>     <chr>       <chr>                                 [1]
#> 1             10 GN 11-434 needles     Tsuga caroliniana Carolina…         0.289
#> 2             11 GN 11-435 needles     Picea abies Norway Spruce           0.288
#> 3             12 GN 11-460 needles     Pinus taeda Loblolly pine           0.322
#> 4             13 HW 07-151 hardwood    Quercus sp. Red oak (from …         0.238
#> 5             14 HW 11-137 hardwood    Acer saccharum Sugar maple          0.243
#> 6             15 HW 11-144 hardwood    Fraxinus americana White a…         0.259
#> # … with 2 more variables: holocellulose [1], spectra <named list>
ir_sample_data[ir_sample_data$sample_type == "office paper", ]
#> # A tibble: 4 × 7
#>   id_measurement id_sample  sample_type  sample_comment klason_lignin
#> *          <int> <chr>      <chr>        <chr>                    [1]
#> 1             55 OFF 10-506 office paper ""                    0.204 
#> 2             56 OFF 13-144 office paper ""                    0.318 
#> 3             57 OFF 08-80  office paper ""                    0.562 
#> 4             58 OFF 08-852 office paper ""                    0.0542
#> # … with 2 more variables: holocellulose [1], spectra <named list>

# subsetting columns
ir_sample_data[, "spectra"]
#> # A tibble: 58 × 1
#>    spectra             
#>  * <named list>        
#>  1 <tibble [3,351 × 2]>
#>  2 <tibble [3,351 × 2]>
#>  3 <tibble [3,351 × 2]>
#>  4 <tibble [3,351 × 2]>
#>  5 <tibble [3,351 × 2]>
#>  6 <tibble [3,351 × 2]>
#>  7 <tibble [3,351 × 2]>
#>  8 <tibble [3,351 × 2]>
#>  9 <tibble [3,351 × 2]>
#> 10 <tibble [3,351 × 2]>
#> # … with 48 more rows
ir_sample_data[["spectra"]]
#> $GN.11.389
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000361
#>  2  3999 0.000431
#>  3  3998 0.000501
#>  4  3997 0.000571
#>  5  3996 0.000667
#>  6  3995 0.000704
#>  7  3994 0.000612
#>  8  3993 0.000525
#>  9  3992 0.000502
#> 10  3991 0.000565
#> # … with 3,341 more rows
#> 
#> $GN.11.400
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000199
#>  2  3999 0.000379
#>  3  3998 0.000558
#>  4  3997 0.000738
#>  5  3996 0.000915
#>  6  3995 0.000987
#>  7  3994 0.000916
#>  8  3993 0.000767
#>  9  3992 0.000658
#> 10  3991 0.000596
#> # … with 3,341 more rows
#> 
#> $GN.11.407
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000104
#>  2  3999 0.000203
#>  3  3998 0.000320
#>  4  3997 0.000394
#>  5  3996 0.000408
#>  6  3995 0.000408
#>  7  3994 0.000452
#>  8  3993 0.000545
#>  9  3992 0.000632
#> 10  3991 0.000650
#> # … with 3,341 more rows
#> 
#> $GN.11.411
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000198
#>  2  3999 0.000231
#>  3  3998 0.000263
#>  4  3997 0.000295
#>  5  3996 0.000341
#>  6  3995 0.000368
#>  7  3994 0.000353
#>  8  3993 0.000297
#>  9  3992 0.000230
#> 10  3991 0.000186
#> # … with 3,341 more rows
#> 
#> $GN.11.416
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000192
#>  2  3999 0.000249
#>  3  3998 0.000277
#>  4  3997 0.000303
#>  5  3996 0.000343
#>  6  3995 0.000397
#>  7  3994 0.000438
#>  8  3993 0.000441
#>  9  3992 0.000405
#> 10  3991 0.000355
#> # … with 3,341 more rows
#> 
#> $GN.11.419
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000228
#>  2  3999 0.000284
#>  3  3998 0.000339
#>  4  3997 0.000394
#>  5  3996 0.000384
#>  6  3995 0.000338
#>  7  3994 0.000295
#>  8  3993 0.000276
#>  9  3992 0.000272
#> 10  3991 0.000267
#> # … with 3,341 more rows
#> 
#> $GN.11.422
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000871
#>  2  3999 0.000139 
#>  3  3998 0.000223 
#>  4  3997 0.000313 
#>  5  3996 0.000375 
#>  6  3995 0.000386 
#>  7  3994 0.000341 
#>  8  3993 0.000258 
#>  9  3992 0.000163 
#> 10  3991 0.0000836
#> # … with 3,341 more rows
#> 
#> $GN.11.423
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000236
#>  2  3999 0.000192
#>  3  3998 0.000147
#>  4  3997 0.000127
#>  5  3996 0.000116
#>  6  3995 0.000143
#>  7  3994 0.000185
#>  8  3993 0.000225
#>  9  3992 0.000242
#> 10  3991 0.000235
#> # … with 3,341 more rows
#> 
#> $GN.11.428
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000111
#>  2  3999 0.000141
#>  3  3998 0.000219
#>  4  3997 0.000323
#>  5  3996 0.000412
#>  6  3995 0.000448
#>  7  3994 0.000421
#>  8  3993 0.000353
#>  9  3992 0.000274
#> 10  3991 0.000205
#> # … with 3,341 more rows
#> 
#> $GN.11.434
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000449
#>  2  3999 0.000388
#>  3  3998 0.000325
#>  4  3997 0.000258
#>  5  3996 0.000226
#>  6  3995 0.000258
#>  7  3994 0.000348
#>  8  3993 0.000461
#>  9  3992 0.000516
#> 10  3991 0.000609
#> # … with 3,341 more rows
#> 
#> $GN.11.435
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000568
#>  2  3999 0.000154 
#>  3  3998 0.000264 
#>  4  3997 0.000352 
#>  5  3996 0.000407 
#>  6  3995 0.000446 
#>  7  3994 0.000494 
#>  8  3993 0.000546 
#>  9  3992 0.000568 
#> 10  3991 0.000524 
#> # … with 3,341 more rows
#> 
#> $GN.11.460
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000385
#>  2  3999 0.000343
#>  3  3998 0.000298
#>  4  3997 0.000250
#>  5  3996 0.000176
#>  6  3995 0.000136
#>  7  3994 0.000143
#>  8  3993 0.000213
#>  9  3992 0.000298
#> 10  3991 0.000364
#> # … with 3,341 more rows
#> 
#> $HW.07.151
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000193 
#>  2  3999 0.000214 
#>  3  3998 0.000179 
#>  4  3997 0.000106 
#>  5  3996 0.0000386
#>  6  3995 0.0000354
#>  7  3994 0.000102 
#>  8  3993 0.000210 
#>  9  3992 0.000294 
#> 10  3991 0.000312 
#> # … with 3,341 more rows
#> 
#> $HW.11.137
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000163 
#>  2  3999 0.000169 
#>  3  3998 0.000142 
#>  4  3997 0.0000846
#>  5  3996 0.0000446
#>  6  3995 0.0000695
#>  7  3994 0.000173 
#>  8  3993 0.000312 
#>  9  3992 0.000409 
#> 10  3991 0.000405 
#> # … with 3,341 more rows
#> 
#> $HW.11.144
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000876
#>  2  3999 0.000194 
#>  3  3998 0.000269 
#>  4  3997 0.000321 
#>  5  3996 0.000342 
#>  6  3995 0.000355 
#>  7  3994 0.000357 
#>  8  3993 0.000326 
#>  9  3992 0.000246 
#> 10  3991 0.000132 
#> # … with 3,341 more rows
#> 
#> $HW.11.146
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000908
#>  2  3999 0.000201 
#>  3  3998 0.000277 
#>  4  3997 0.000297 
#>  5  3996 0.000296 
#>  6  3995 0.000287 
#>  7  3994 0.000302 
#>  8  3993 0.000327 
#>  9  3992 0.000321 
#> 10  3991 0.000261 
#> # … with 3,341 more rows
#> 
#> $HW.11.148
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000196 
#>  2  3999 0.000182 
#>  3  3998 0.000146 
#>  4  3997 0.0000904
#>  5  3996 0.0000709
#>  6  3995 0.000118 
#>  7  3994 0.000211 
#>  8  3993 0.000294 
#>  9  3992 0.000316 
#> 10  3991 0.000272 
#> # … with 3,341 more rows
#> 
#> $HW.11.149
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000203 
#>  2  3999 0.000254 
#>  3  3998 0.000224 
#>  4  3997 0.000141 
#>  5  3996 0.0000606
#>  6  3995 0.0000288
#>  7  3994 0.000055 
#>  8  3993 0.000112 
#>  9  3992 0.000143 
#> 10  3991 0.000129 
#> # … with 3,341 more rows
#> 
#> $HW.11.172
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000164
#>  2  3999 0.000215
#>  3  3998 0.000240
#>  4  3997 0.000223
#>  5  3996 0.000193
#>  6  3995 0.000196
#>  7  3994 0.000248
#>  8  3993 0.000310
#>  9  3992 0.000321
#> 10  3991 0.000256
#> # … with 3,341 more rows
#> 
#> $HW.11.173
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000134 
#>  2  3999 0.000210 
#>  3  3998 0.000221 
#>  4  3997 0.000155 
#>  5  3996 0.0000658
#>  6  3995 0.0000434
#>  7  3994 0.0000766
#>  8  3993 0.000155 
#>  9  3992 0.000213 
#> 10  3991 0.000203 
#> # … with 3,341 more rows
#> 
#> $HW.11.176
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000143 
#>  2  3999 0.000150 
#>  3  3998 0.000113 
#>  4  3997 0.0000535
#>  5  3996 0.0000051
#>  6  3995 0        
#>  7  3994 0.0000455
#>  8  3993 0.000117 
#>  9  3992 0.000175 
#> 10  3991 0.000195 
#> # … with 3,341 more rows
#> 
#> $HW.11.179
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000197 
#>  2  3999 0.000177 
#>  3  3998 0.000140 
#>  4  3997 0.000102 
#>  5  3996 0.0000988
#>  6  3995 0.000157 
#>  7  3994 0.000261 
#>  8  3993 0.000360 
#>  9  3992 0.000394 
#> 10  3991 0.000340 
#> # … with 3,341 more rows
#> 
#> $HW.11.184
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000136
#>  2  3999 0.000222
#>  3  3998 0.000274
#>  4  3997 0.000271
#>  5  3996 0.000209
#>  6  3995 0.000143
#>  7  3994 0.000119
#>  8  3993 0.000139
#>  9  3992 0.000166
#> 10  3991 0.000154
#> # … with 3,341 more rows
#> 
#> $LG.09.384
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000141
#>  2  3999 0.000105
#>  3  3998 0.000111
#>  4  3997 0.000146
#>  5  3996 0.000181
#>  6  3995 0.000204
#>  7  3994 0.000212
#>  8  3993 0.000207
#>  9  3992 0.000192
#> 10  3991 0.000180
#> # … with 3,341 more rows
#> 
#> $LG.11.309
#> # A tibble: 3,351 × 2
#>        x          y
#>    <int>      <dbl>
#>  1  4000 0.000167  
#>  2  3999 0.000238  
#>  3  3998 0.000270  
#>  4  3997 0.000243  
#>  5  3996 0.000178  
#>  6  3995 0.0000299 
#>  7  3994 0.00000686
#>  8  3993 0.0000312 
#>  9  3992 0.0000693 
#> 10  3991 0.0000924 
#> # … with 3,341 more rows
#> 
#> $LG.11.393
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000608
#>  2  3999 0.000618
#>  3  3998 0.000655
#>  4  3997 0.000697
#>  5  3996 0.000732
#>  6  3995 0.000755
#>  7  3994 0.000762
#>  8  3993 0.000748
#>  9  3992 0.000720
#> 10  3991 0.000702
#> # … with 3,341 more rows
#> 
#> $LG.11.395
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000138
#>  2  3999 0.000274
#>  3  3998 0.000436
#>  4  3997 0.000569
#>  5  3996 0.000630
#>  6  3995 0.000625
#>  7  3994 0.000595
#>  8  3993 0.000574
#>  9  3992 0.000564
#> 10  3991 0.000553
#> # … with 3,341 more rows
#> 
#> $LG.11.397
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000229
#>  2  3999 0.000342
#>  3  3998 0.000404
#>  4  3997 0.000436
#>  5  3996 0.000462
#>  6  3995 0.000500
#>  7  3994 0.000539
#>  8  3993 0.000552
#>  9  3992 0.000531
#> 10  3991 0.000495
#> # … with 3,341 more rows
#> 
#> $LG.11.404
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000165
#>  2  3999 0.000256
#>  3  3998 0.000347
#>  4  3997 0.000369
#>  5  3996 0.000332
#>  6  3995 0.000275
#>  7  3994 0.000235
#>  8  3993 0.000221
#>  9  3992 0.000224
#> 10  3991 0.000234
#> # … with 3,341 more rows
#> 
#> $LG.11.412
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000087 
#>  2  3999 0.000270 
#>  3  3998 0.000231 
#>  4  3997 0.000263 
#>  5  3996 0.000222 
#>  6  3995 0.000148 
#>  7  3994 0.0000888
#>  8  3993 0.0000699
#>  9  3992 0.0000886
#> 10  3991 0.000130 
#> # … with 3,341 more rows
#> 
#> $LG.11.414
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000105
#>  2  3999 0.000166
#>  3  3998 0.000203
#>  4  3997 0.000227
#>  5  3996 0.000251
#>  6  3995 0.000276
#>  7  3994 0.000284
#>  8  3993 0.000262
#>  9  3992 0.000219
#> 10  3991 0.000183
#> # … with 3,341 more rows
#> 
#> $LG.11.418
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000367
#>  2  3999 0.000355
#>  3  3998 0.000342
#>  4  3997 0.000295
#>  5  3996 0.000248
#>  6  3995 0.000233
#>  7  3994 0.000249
#>  8  3993 0.000269
#>  9  3992 0.000278
#> 10  3991 0.000288
#> # … with 3,341 more rows
#> 
#> $LG.11.432
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000238
#>  2  3999 0.000353
#>  3  3998 0.000443
#>  4  3997 0.000494
#>  5  3996 0.000510
#>  6  3995 0.000519
#>  7  3994 0.000535
#>  8  3993 0.000548
#>  9  3992 0.000544
#> 10  3991 0.000527
#> # … with 3,341 more rows
#> 
#> $LG.11.433
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000626
#>  2  3999 0.000638
#>  3  3998 0.000652
#>  4  3997 0.000510
#>  5  3996 0.000493
#>  6  3995 0.000505
#>  7  3994 0.000540
#>  8  3993 0.000577
#>  9  3992 0.000598
#> 10  3991 0.000601
#> # … with 3,341 more rows
#> 
#> $LG.11.438
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000477
#>  2  3999 0.000422
#>  3  3998 0.000390
#>  4  3997 0.000384
#>  5  3996 0.000383
#>  6  3995 0.000379
#>  7  3994 0.000383
#>  8  3993 0.000411
#>  9  3992 0.000462
#> 10  3991 0.000519
#> # … with 3,341 more rows
#> 
#> $LG.11.461
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000813
#>  2  3999 0.000152 
#>  3  3998 0.000223 
#>  4  3997 0.000276 
#>  5  3996 0.000298 
#>  6  3995 0.000297 
#>  7  3994 0.000285 
#>  8  3993 0.000272 
#>  9  3992 0.000255 
#> 10  3991 0.000238 
#> # … with 3,341 more rows
#> 
#> $SW.11.138
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000125 
#>  2  3999 0.000150 
#>  3  3998 0.000128 
#>  4  3997 0.0000665
#>  5  3996 0.0000404
#>  6  3995 0.0000627
#>  7  3994 0.000152 
#>  8  3993 0.000253 
#>  9  3992 0.000301 
#> 10  3991 0.000267 
#> # … with 3,341 more rows
#> 
#> $SW.11.139
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000183
#>  2  3999 0.000265
#>  3  3998 0.000290
#>  4  3997 0.000255
#>  5  3996 0.000192
#>  6  3995 0.000142
#>  7  3994 0.000152
#>  8  3993 0.000152
#>  9  3992 0.000139
#> 10  3991 0.000103
#> # … with 3,341 more rows
#> 
#> $SW.11.154
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000175
#>  2  3999 0.000258
#>  3  3998 0.000285
#>  4  3997 0.000250
#>  5  3996 0.000228
#>  6  3995 0.000214
#>  7  3994 0.000243
#>  8  3993 0.000270
#>  9  3992 0.000242
#> 10  3991 0.000155
#> # … with 3,341 more rows
#> 
#> $SW.11.158
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000156
#>  2  3999 0.000241
#>  3  3998 0.000254
#>  4  3997 0.000201
#>  5  3996 0.000139
#>  6  3995 0.000155
#>  7  3994 0.000204
#>  8  3993 0.000266
#>  9  3992 0.000276
#> 10  3991 0.000209
#> # … with 3,341 more rows
#> 
#> $SW.11.178
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000212
#>  2  3999 0.000234
#>  3  3998 0.000200
#>  4  3997 0.000141
#>  5  3996 0.000105
#>  6  3995 0.000128
#>  7  3994 0.000204
#>  8  3993 0.000284
#>  9  3992 0.000303
#> 10  3991 0.000235
#> # … with 3,341 more rows
#> 
#> $OCC.11.457
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000155
#>  2  3999 0.000189
#>  3  3998 0.000216
#>  4  3997 0.000236
#>  5  3996 0.000245
#>  6  3995 0.000255
#>  7  3994 0.000278
#>  8  3993 0.000313
#>  9  3992 0.000330
#> 10  3991 0.000293
#> # … with 3,341 more rows
#> 
#> $OCC.11.462
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000218 
#>  2  3999 0.000210 
#>  3  3998 0.000151 
#>  4  3997 0.0000726
#>  5  3996 0.0000238
#>  6  3995 0.0000902
#>  7  3994 0.000158 
#>  8  3993 0.000232 
#>  9  3992 0.000261 
#> 10  3991 0.000227 
#> # … with 3,341 more rows
#> 
#> $OCC.11.463
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000111
#>  2  3999 0.000247
#>  3  3998 0.000328
#>  4  3997 0.000360
#>  5  3996 0.000317
#>  6  3995 0.000284
#>  7  3994 0.000296
#>  8  3993 0.000329
#>  9  3992 0.000318
#> 10  3991 0.000224
#> # … with 3,341 more rows
#> 
#> $OMG.08.82
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000227
#>  2  3999 0.000259
#>  3  3998 0.000301
#>  4  3997 0.000345
#>  5  3996 0.000403
#>  6  3995 0.000464
#>  7  3994 0.000505
#>  8  3993 0.000501
#>  9  3992 0.000456
#> 10  3991 0.000403
#> # … with 3,341 more rows
#> 
#> $OMG.11.454
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000224
#>  2  3999 0.000263
#>  3  3998 0.000254
#>  4  3997 0.000202
#>  5  3996 0.000140
#>  6  3995 0.000105
#>  7  3994 0.000117
#>  8  3993 0.000167
#>  9  3992 0.000237
#> 10  3991 0.000309
#> # … with 3,341 more rows
#> 
#> $OMG.11.455
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000248
#>  2  3999 0.000238
#>  3  3998 0.000227
#>  4  3997 0.000216
#>  5  3996 0.000204
#>  6  3995 0.000191
#>  7  3994 0.000153
#>  8  3993 0.000124
#>  9  3992 0.000138
#> 10  3991 0.000196
#> # … with 3,341 more rows
#> 
#> $OMG.11.456
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000300
#>  2  3999 0.000262
#>  3  3998 0.000222
#>  4  3997 0.000179
#>  5  3996 0.000160
#>  6  3995 0.000145
#>  7  3994 0.000125
#>  8  3993 0.000119
#>  9  3992 0.000160
#> 10  3991 0.000264
#> # … with 3,341 more rows
#> 
#> $ONP.08.78
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000215 
#>  2  3999 0.000189 
#>  3  3998 0.000149 
#>  4  3997 0.000109 
#>  5  3996 0.0000816
#>  6  3995 0.0000732
#>  7  3994 0.000083 
#>  8  3993 0.000106 
#>  9  3992 0.000138 
#> 10  3991 0.000182 
#> # … with 3,341 more rows
#> 
#> $ONP.09.388
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000260
#>  2  3999 0.000193
#>  3  3998 0.000137
#>  4  3997 0.000118
#>  5  3996 0.000137
#>  6  3995 0.000174
#>  7  3994 0.000211
#>  8  3993 0.000239
#>  9  3992 0.000267
#> 10  3991 0.000311
#> # … with 3,341 more rows
#> 
#> $ONP.11.450
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000255
#>  2  3999 0.000261
#>  3  3998 0.000268
#>  4  3997 0.000274
#>  5  3996 0.000316
#>  6  3995 0.000350
#>  7  3994 0.000349
#>  8  3993 0.000328
#>  9  3992 0.000330
#> 10  3991 0.000389
#> # … with 3,341 more rows
#> 
#> $ONP.11.451
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000244
#>  2  3999 0.000244
#>  3  3998 0.000244
#>  4  3997 0.000243
#>  5  3996 0.000261
#>  6  3995 0.000277
#>  7  3994 0.000277
#>  8  3993 0.000263
#>  9  3992 0.000255
#> 10  3991 0.000271
#> # … with 3,341 more rows
#> 
#> $ONP.11.458
#> # A tibble: 3,351 × 2
#>        x          y
#>    <int>      <dbl>
#>  1  4000 0.000129  
#>  2  3999 0.0000484 
#>  3  3998 0         
#>  4  3997 0.00000254
#>  5  3996 0.0000362 
#>  6  3995 0.0000648 
#>  7  3994 0.0000706 
#>  8  3993 0.0000684 
#>  9  3992 0.0000893 
#> 10  3991 0.000152  
#> # … with 3,341 more rows
#> 
#> $ONP.11.459
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000235 
#>  2  3999 0.000217 
#>  3  3998 0.000198 
#>  4  3997 0.000177 
#>  5  3996 0.000154 
#>  6  3995 0.000122 
#>  7  3994 0.0000991
#>  8  3993 0.000110 
#>  9  3992 0.000164 
#> 10  3991 0.000249 
#> # … with 3,341 more rows
#> 
#> $OFF.10.506
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000281
#>  2  3999 0.000313
#>  3  3998 0.000283
#>  4  3997 0.000213
#>  5  3996 0.000153
#>  6  3995 0.000147
#>  7  3994 0.000199
#>  8  3993 0.000265
#>  9  3992 0.000292
#> 10  3991 0.000255
#> # … with 3,341 more rows
#> 
#> $OFF.13.144
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000118
#>  2  3999 0.000253
#>  3  3998 0.000338
#>  4  3997 0.000337
#>  5  3996 0.000280
#>  6  3995 0.000222
#>  7  3994 0.000231
#>  8  3993 0.000219
#>  9  3992 0.000183
#> 10  3991 0.000116
#> # … with 3,341 more rows
#> 
#> $OFF.08.80
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000170
#>  2  3999 0.000212
#>  3  3998 0.000194
#>  4  3997 0.000135
#>  5  3996 0.000124
#>  6  3995 0.000159
#>  7  3994 0.000256
#>  8  3993 0.000352
#>  9  3992 0.000379
#> 10  3991 0.000314
#> # … with 3,341 more rows
#> 
#> $OFF.08.852
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000108
#>  2  3999 0.000186
#>  3  3998 0.000209
#>  4  3997 0.000177
#>  5  3996 0.000136
#>  6  3995 0.000177
#>  7  3994 0.000231
#>  8  3993 0.000303
#>  9  3992 0.000338
#> 10  3991 0.000304
#> # … with 3,341 more rows
#> 
#> attr(,"class")
#> [1] "list"
ir_sample_data$spectra
#> $GN.11.389
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000361
#>  2  3999 0.000431
#>  3  3998 0.000501
#>  4  3997 0.000571
#>  5  3996 0.000667
#>  6  3995 0.000704
#>  7  3994 0.000612
#>  8  3993 0.000525
#>  9  3992 0.000502
#> 10  3991 0.000565
#> # … with 3,341 more rows
#> 
#> $GN.11.400
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000199
#>  2  3999 0.000379
#>  3  3998 0.000558
#>  4  3997 0.000738
#>  5  3996 0.000915
#>  6  3995 0.000987
#>  7  3994 0.000916
#>  8  3993 0.000767
#>  9  3992 0.000658
#> 10  3991 0.000596
#> # … with 3,341 more rows
#> 
#> $GN.11.407
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000104
#>  2  3999 0.000203
#>  3  3998 0.000320
#>  4  3997 0.000394
#>  5  3996 0.000408
#>  6  3995 0.000408
#>  7  3994 0.000452
#>  8  3993 0.000545
#>  9  3992 0.000632
#> 10  3991 0.000650
#> # … with 3,341 more rows
#> 
#> $GN.11.411
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000198
#>  2  3999 0.000231
#>  3  3998 0.000263
#>  4  3997 0.000295
#>  5  3996 0.000341
#>  6  3995 0.000368
#>  7  3994 0.000353
#>  8  3993 0.000297
#>  9  3992 0.000230
#> 10  3991 0.000186
#> # … with 3,341 more rows
#> 
#> $GN.11.416
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000192
#>  2  3999 0.000249
#>  3  3998 0.000277
#>  4  3997 0.000303
#>  5  3996 0.000343
#>  6  3995 0.000397
#>  7  3994 0.000438
#>  8  3993 0.000441
#>  9  3992 0.000405
#> 10  3991 0.000355
#> # … with 3,341 more rows
#> 
#> $GN.11.419
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000228
#>  2  3999 0.000284
#>  3  3998 0.000339
#>  4  3997 0.000394
#>  5  3996 0.000384
#>  6  3995 0.000338
#>  7  3994 0.000295
#>  8  3993 0.000276
#>  9  3992 0.000272
#> 10  3991 0.000267
#> # … with 3,341 more rows
#> 
#> $GN.11.422
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000871
#>  2  3999 0.000139 
#>  3  3998 0.000223 
#>  4  3997 0.000313 
#>  5  3996 0.000375 
#>  6  3995 0.000386 
#>  7  3994 0.000341 
#>  8  3993 0.000258 
#>  9  3992 0.000163 
#> 10  3991 0.0000836
#> # … with 3,341 more rows
#> 
#> $GN.11.423
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000236
#>  2  3999 0.000192
#>  3  3998 0.000147
#>  4  3997 0.000127
#>  5  3996 0.000116
#>  6  3995 0.000143
#>  7  3994 0.000185
#>  8  3993 0.000225
#>  9  3992 0.000242
#> 10  3991 0.000235
#> # … with 3,341 more rows
#> 
#> $GN.11.428
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000111
#>  2  3999 0.000141
#>  3  3998 0.000219
#>  4  3997 0.000323
#>  5  3996 0.000412
#>  6  3995 0.000448
#>  7  3994 0.000421
#>  8  3993 0.000353
#>  9  3992 0.000274
#> 10  3991 0.000205
#> # … with 3,341 more rows
#> 
#> $GN.11.434
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000449
#>  2  3999 0.000388
#>  3  3998 0.000325
#>  4  3997 0.000258
#>  5  3996 0.000226
#>  6  3995 0.000258
#>  7  3994 0.000348
#>  8  3993 0.000461
#>  9  3992 0.000516
#> 10  3991 0.000609
#> # … with 3,341 more rows
#> 
#> $GN.11.435
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000568
#>  2  3999 0.000154 
#>  3  3998 0.000264 
#>  4  3997 0.000352 
#>  5  3996 0.000407 
#>  6  3995 0.000446 
#>  7  3994 0.000494 
#>  8  3993 0.000546 
#>  9  3992 0.000568 
#> 10  3991 0.000524 
#> # … with 3,341 more rows
#> 
#> $GN.11.460
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000385
#>  2  3999 0.000343
#>  3  3998 0.000298
#>  4  3997 0.000250
#>  5  3996 0.000176
#>  6  3995 0.000136
#>  7  3994 0.000143
#>  8  3993 0.000213
#>  9  3992 0.000298
#> 10  3991 0.000364
#> # … with 3,341 more rows
#> 
#> $HW.07.151
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000193 
#>  2  3999 0.000214 
#>  3  3998 0.000179 
#>  4  3997 0.000106 
#>  5  3996 0.0000386
#>  6  3995 0.0000354
#>  7  3994 0.000102 
#>  8  3993 0.000210 
#>  9  3992 0.000294 
#> 10  3991 0.000312 
#> # … with 3,341 more rows
#> 
#> $HW.11.137
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000163 
#>  2  3999 0.000169 
#>  3  3998 0.000142 
#>  4  3997 0.0000846
#>  5  3996 0.0000446
#>  6  3995 0.0000695
#>  7  3994 0.000173 
#>  8  3993 0.000312 
#>  9  3992 0.000409 
#> 10  3991 0.000405 
#> # … with 3,341 more rows
#> 
#> $HW.11.144
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000876
#>  2  3999 0.000194 
#>  3  3998 0.000269 
#>  4  3997 0.000321 
#>  5  3996 0.000342 
#>  6  3995 0.000355 
#>  7  3994 0.000357 
#>  8  3993 0.000326 
#>  9  3992 0.000246 
#> 10  3991 0.000132 
#> # … with 3,341 more rows
#> 
#> $HW.11.146
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000908
#>  2  3999 0.000201 
#>  3  3998 0.000277 
#>  4  3997 0.000297 
#>  5  3996 0.000296 
#>  6  3995 0.000287 
#>  7  3994 0.000302 
#>  8  3993 0.000327 
#>  9  3992 0.000321 
#> 10  3991 0.000261 
#> # … with 3,341 more rows
#> 
#> $HW.11.148
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000196 
#>  2  3999 0.000182 
#>  3  3998 0.000146 
#>  4  3997 0.0000904
#>  5  3996 0.0000709
#>  6  3995 0.000118 
#>  7  3994 0.000211 
#>  8  3993 0.000294 
#>  9  3992 0.000316 
#> 10  3991 0.000272 
#> # … with 3,341 more rows
#> 
#> $HW.11.149
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000203 
#>  2  3999 0.000254 
#>  3  3998 0.000224 
#>  4  3997 0.000141 
#>  5  3996 0.0000606
#>  6  3995 0.0000288
#>  7  3994 0.000055 
#>  8  3993 0.000112 
#>  9  3992 0.000143 
#> 10  3991 0.000129 
#> # … with 3,341 more rows
#> 
#> $HW.11.172
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000164
#>  2  3999 0.000215
#>  3  3998 0.000240
#>  4  3997 0.000223
#>  5  3996 0.000193
#>  6  3995 0.000196
#>  7  3994 0.000248
#>  8  3993 0.000310
#>  9  3992 0.000321
#> 10  3991 0.000256
#> # … with 3,341 more rows
#> 
#> $HW.11.173
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000134 
#>  2  3999 0.000210 
#>  3  3998 0.000221 
#>  4  3997 0.000155 
#>  5  3996 0.0000658
#>  6  3995 0.0000434
#>  7  3994 0.0000766
#>  8  3993 0.000155 
#>  9  3992 0.000213 
#> 10  3991 0.000203 
#> # … with 3,341 more rows
#> 
#> $HW.11.176
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000143 
#>  2  3999 0.000150 
#>  3  3998 0.000113 
#>  4  3997 0.0000535
#>  5  3996 0.0000051
#>  6  3995 0        
#>  7  3994 0.0000455
#>  8  3993 0.000117 
#>  9  3992 0.000175 
#> 10  3991 0.000195 
#> # … with 3,341 more rows
#> 
#> $HW.11.179
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000197 
#>  2  3999 0.000177 
#>  3  3998 0.000140 
#>  4  3997 0.000102 
#>  5  3996 0.0000988
#>  6  3995 0.000157 
#>  7  3994 0.000261 
#>  8  3993 0.000360 
#>  9  3992 0.000394 
#> 10  3991 0.000340 
#> # … with 3,341 more rows
#> 
#> $HW.11.184
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000136
#>  2  3999 0.000222
#>  3  3998 0.000274
#>  4  3997 0.000271
#>  5  3996 0.000209
#>  6  3995 0.000143
#>  7  3994 0.000119
#>  8  3993 0.000139
#>  9  3992 0.000166
#> 10  3991 0.000154
#> # … with 3,341 more rows
#> 
#> $LG.09.384
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000141
#>  2  3999 0.000105
#>  3  3998 0.000111
#>  4  3997 0.000146
#>  5  3996 0.000181
#>  6  3995 0.000204
#>  7  3994 0.000212
#>  8  3993 0.000207
#>  9  3992 0.000192
#> 10  3991 0.000180
#> # … with 3,341 more rows
#> 
#> $LG.11.309
#> # A tibble: 3,351 × 2
#>        x          y
#>    <int>      <dbl>
#>  1  4000 0.000167  
#>  2  3999 0.000238  
#>  3  3998 0.000270  
#>  4  3997 0.000243  
#>  5  3996 0.000178  
#>  6  3995 0.0000299 
#>  7  3994 0.00000686
#>  8  3993 0.0000312 
#>  9  3992 0.0000693 
#> 10  3991 0.0000924 
#> # … with 3,341 more rows
#> 
#> $LG.11.393
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000608
#>  2  3999 0.000618
#>  3  3998 0.000655
#>  4  3997 0.000697
#>  5  3996 0.000732
#>  6  3995 0.000755
#>  7  3994 0.000762
#>  8  3993 0.000748
#>  9  3992 0.000720
#> 10  3991 0.000702
#> # … with 3,341 more rows
#> 
#> $LG.11.395
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000138
#>  2  3999 0.000274
#>  3  3998 0.000436
#>  4  3997 0.000569
#>  5  3996 0.000630
#>  6  3995 0.000625
#>  7  3994 0.000595
#>  8  3993 0.000574
#>  9  3992 0.000564
#> 10  3991 0.000553
#> # … with 3,341 more rows
#> 
#> $LG.11.397
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000229
#>  2  3999 0.000342
#>  3  3998 0.000404
#>  4  3997 0.000436
#>  5  3996 0.000462
#>  6  3995 0.000500
#>  7  3994 0.000539
#>  8  3993 0.000552
#>  9  3992 0.000531
#> 10  3991 0.000495
#> # … with 3,341 more rows
#> 
#> $LG.11.404
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000165
#>  2  3999 0.000256
#>  3  3998 0.000347
#>  4  3997 0.000369
#>  5  3996 0.000332
#>  6  3995 0.000275
#>  7  3994 0.000235
#>  8  3993 0.000221
#>  9  3992 0.000224
#> 10  3991 0.000234
#> # … with 3,341 more rows
#> 
#> $LG.11.412
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000087 
#>  2  3999 0.000270 
#>  3  3998 0.000231 
#>  4  3997 0.000263 
#>  5  3996 0.000222 
#>  6  3995 0.000148 
#>  7  3994 0.0000888
#>  8  3993 0.0000699
#>  9  3992 0.0000886
#> 10  3991 0.000130 
#> # … with 3,341 more rows
#> 
#> $LG.11.414
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000105
#>  2  3999 0.000166
#>  3  3998 0.000203
#>  4  3997 0.000227
#>  5  3996 0.000251
#>  6  3995 0.000276
#>  7  3994 0.000284
#>  8  3993 0.000262
#>  9  3992 0.000219
#> 10  3991 0.000183
#> # … with 3,341 more rows
#> 
#> $LG.11.418
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000367
#>  2  3999 0.000355
#>  3  3998 0.000342
#>  4  3997 0.000295
#>  5  3996 0.000248
#>  6  3995 0.000233
#>  7  3994 0.000249
#>  8  3993 0.000269
#>  9  3992 0.000278
#> 10  3991 0.000288
#> # … with 3,341 more rows
#> 
#> $LG.11.432
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000238
#>  2  3999 0.000353
#>  3  3998 0.000443
#>  4  3997 0.000494
#>  5  3996 0.000510
#>  6  3995 0.000519
#>  7  3994 0.000535
#>  8  3993 0.000548
#>  9  3992 0.000544
#> 10  3991 0.000527
#> # … with 3,341 more rows
#> 
#> $LG.11.433
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000626
#>  2  3999 0.000638
#>  3  3998 0.000652
#>  4  3997 0.000510
#>  5  3996 0.000493
#>  6  3995 0.000505
#>  7  3994 0.000540
#>  8  3993 0.000577
#>  9  3992 0.000598
#> 10  3991 0.000601
#> # … with 3,341 more rows
#> 
#> $LG.11.438
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000477
#>  2  3999 0.000422
#>  3  3998 0.000390
#>  4  3997 0.000384
#>  5  3996 0.000383
#>  6  3995 0.000379
#>  7  3994 0.000383
#>  8  3993 0.000411
#>  9  3992 0.000462
#> 10  3991 0.000519
#> # … with 3,341 more rows
#> 
#> $LG.11.461
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.0000813
#>  2  3999 0.000152 
#>  3  3998 0.000223 
#>  4  3997 0.000276 
#>  5  3996 0.000298 
#>  6  3995 0.000297 
#>  7  3994 0.000285 
#>  8  3993 0.000272 
#>  9  3992 0.000255 
#> 10  3991 0.000238 
#> # … with 3,341 more rows
#> 
#> $SW.11.138
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000125 
#>  2  3999 0.000150 
#>  3  3998 0.000128 
#>  4  3997 0.0000665
#>  5  3996 0.0000404
#>  6  3995 0.0000627
#>  7  3994 0.000152 
#>  8  3993 0.000253 
#>  9  3992 0.000301 
#> 10  3991 0.000267 
#> # … with 3,341 more rows
#> 
#> $SW.11.139
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000183
#>  2  3999 0.000265
#>  3  3998 0.000290
#>  4  3997 0.000255
#>  5  3996 0.000192
#>  6  3995 0.000142
#>  7  3994 0.000152
#>  8  3993 0.000152
#>  9  3992 0.000139
#> 10  3991 0.000103
#> # … with 3,341 more rows
#> 
#> $SW.11.154
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000175
#>  2  3999 0.000258
#>  3  3998 0.000285
#>  4  3997 0.000250
#>  5  3996 0.000228
#>  6  3995 0.000214
#>  7  3994 0.000243
#>  8  3993 0.000270
#>  9  3992 0.000242
#> 10  3991 0.000155
#> # … with 3,341 more rows
#> 
#> $SW.11.158
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000156
#>  2  3999 0.000241
#>  3  3998 0.000254
#>  4  3997 0.000201
#>  5  3996 0.000139
#>  6  3995 0.000155
#>  7  3994 0.000204
#>  8  3993 0.000266
#>  9  3992 0.000276
#> 10  3991 0.000209
#> # … with 3,341 more rows
#> 
#> $SW.11.178
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000212
#>  2  3999 0.000234
#>  3  3998 0.000200
#>  4  3997 0.000141
#>  5  3996 0.000105
#>  6  3995 0.000128
#>  7  3994 0.000204
#>  8  3993 0.000284
#>  9  3992 0.000303
#> 10  3991 0.000235
#> # … with 3,341 more rows
#> 
#> $OCC.11.457
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000155
#>  2  3999 0.000189
#>  3  3998 0.000216
#>  4  3997 0.000236
#>  5  3996 0.000245
#>  6  3995 0.000255
#>  7  3994 0.000278
#>  8  3993 0.000313
#>  9  3992 0.000330
#> 10  3991 0.000293
#> # … with 3,341 more rows
#> 
#> $OCC.11.462
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000218 
#>  2  3999 0.000210 
#>  3  3998 0.000151 
#>  4  3997 0.0000726
#>  5  3996 0.0000238
#>  6  3995 0.0000902
#>  7  3994 0.000158 
#>  8  3993 0.000232 
#>  9  3992 0.000261 
#> 10  3991 0.000227 
#> # … with 3,341 more rows
#> 
#> $OCC.11.463
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000111
#>  2  3999 0.000247
#>  3  3998 0.000328
#>  4  3997 0.000360
#>  5  3996 0.000317
#>  6  3995 0.000284
#>  7  3994 0.000296
#>  8  3993 0.000329
#>  9  3992 0.000318
#> 10  3991 0.000224
#> # … with 3,341 more rows
#> 
#> $OMG.08.82
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000227
#>  2  3999 0.000259
#>  3  3998 0.000301
#>  4  3997 0.000345
#>  5  3996 0.000403
#>  6  3995 0.000464
#>  7  3994 0.000505
#>  8  3993 0.000501
#>  9  3992 0.000456
#> 10  3991 0.000403
#> # … with 3,341 more rows
#> 
#> $OMG.11.454
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000224
#>  2  3999 0.000263
#>  3  3998 0.000254
#>  4  3997 0.000202
#>  5  3996 0.000140
#>  6  3995 0.000105
#>  7  3994 0.000117
#>  8  3993 0.000167
#>  9  3992 0.000237
#> 10  3991 0.000309
#> # … with 3,341 more rows
#> 
#> $OMG.11.455
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000248
#>  2  3999 0.000238
#>  3  3998 0.000227
#>  4  3997 0.000216
#>  5  3996 0.000204
#>  6  3995 0.000191
#>  7  3994 0.000153
#>  8  3993 0.000124
#>  9  3992 0.000138
#> 10  3991 0.000196
#> # … with 3,341 more rows
#> 
#> $OMG.11.456
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000300
#>  2  3999 0.000262
#>  3  3998 0.000222
#>  4  3997 0.000179
#>  5  3996 0.000160
#>  6  3995 0.000145
#>  7  3994 0.000125
#>  8  3993 0.000119
#>  9  3992 0.000160
#> 10  3991 0.000264
#> # … with 3,341 more rows
#> 
#> $ONP.08.78
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000215 
#>  2  3999 0.000189 
#>  3  3998 0.000149 
#>  4  3997 0.000109 
#>  5  3996 0.0000816
#>  6  3995 0.0000732
#>  7  3994 0.000083 
#>  8  3993 0.000106 
#>  9  3992 0.000138 
#> 10  3991 0.000182 
#> # … with 3,341 more rows
#> 
#> $ONP.09.388
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000260
#>  2  3999 0.000193
#>  3  3998 0.000137
#>  4  3997 0.000118
#>  5  3996 0.000137
#>  6  3995 0.000174
#>  7  3994 0.000211
#>  8  3993 0.000239
#>  9  3992 0.000267
#> 10  3991 0.000311
#> # … with 3,341 more rows
#> 
#> $ONP.11.450
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000255
#>  2  3999 0.000261
#>  3  3998 0.000268
#>  4  3997 0.000274
#>  5  3996 0.000316
#>  6  3995 0.000350
#>  7  3994 0.000349
#>  8  3993 0.000328
#>  9  3992 0.000330
#> 10  3991 0.000389
#> # … with 3,341 more rows
#> 
#> $ONP.11.451
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000244
#>  2  3999 0.000244
#>  3  3998 0.000244
#>  4  3997 0.000243
#>  5  3996 0.000261
#>  6  3995 0.000277
#>  7  3994 0.000277
#>  8  3993 0.000263
#>  9  3992 0.000255
#> 10  3991 0.000271
#> # … with 3,341 more rows
#> 
#> $ONP.11.458
#> # A tibble: 3,351 × 2
#>        x          y
#>    <int>      <dbl>
#>  1  4000 0.000129  
#>  2  3999 0.0000484 
#>  3  3998 0         
#>  4  3997 0.00000254
#>  5  3996 0.0000362 
#>  6  3995 0.0000648 
#>  7  3994 0.0000706 
#>  8  3993 0.0000684 
#>  9  3992 0.0000893 
#> 10  3991 0.000152  
#> # … with 3,341 more rows
#> 
#> $ONP.11.459
#> # A tibble: 3,351 × 2
#>        x         y
#>    <int>     <dbl>
#>  1  4000 0.000235 
#>  2  3999 0.000217 
#>  3  3998 0.000198 
#>  4  3997 0.000177 
#>  5  3996 0.000154 
#>  6  3995 0.000122 
#>  7  3994 0.0000991
#>  8  3993 0.000110 
#>  9  3992 0.000164 
#> 10  3991 0.000249 
#> # … with 3,341 more rows
#> 
#> $OFF.10.506
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000281
#>  2  3999 0.000313
#>  3  3998 0.000283
#>  4  3997 0.000213
#>  5  3996 0.000153
#>  6  3995 0.000147
#>  7  3994 0.000199
#>  8  3993 0.000265
#>  9  3992 0.000292
#> 10  3991 0.000255
#> # … with 3,341 more rows
#> 
#> $OFF.13.144
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000118
#>  2  3999 0.000253
#>  3  3998 0.000338
#>  4  3997 0.000337
#>  5  3996 0.000280
#>  6  3995 0.000222
#>  7  3994 0.000231
#>  8  3993 0.000219
#>  9  3992 0.000183
#> 10  3991 0.000116
#> # … with 3,341 more rows
#> 
#> $OFF.08.80
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000170
#>  2  3999 0.000212
#>  3  3998 0.000194
#>  4  3997 0.000135
#>  5  3996 0.000124
#>  6  3995 0.000159
#>  7  3994 0.000256
#>  8  3993 0.000352
#>  9  3992 0.000379
#> 10  3991 0.000314
#> # … with 3,341 more rows
#> 
#> $OFF.08.852
#> # A tibble: 3,351 × 2
#>        x        y
#>    <int>    <dbl>
#>  1  4000 0.000108
#>  2  3999 0.000186
#>  3  3998 0.000209
#>  4  3997 0.000177
#>  5  3996 0.000136
#>  6  3995 0.000177
#>  7  3994 0.000231
#>  8  3993 0.000303
#>  9  3992 0.000338
#> 10  3991 0.000304
#> # … with 3,341 more rows
#> 
#> attr(,"class")
#> [1] "list"

# not explicitly selecting the spectra column drops the ir class
class(ir_sample_data[, 1])
#> [1] "tbl_df"     "tbl"        "data.frame"
class(ir_sample_data[, "spectra"])
#> [1] "ir"         "tbl_df"     "tbl"        "data.frame"

# subsetting values
ir_sample_data[, 1] # drops the ir class
#> # A tibble: 58 × 1
#>    id_measurement
#>  *          <int>
#>  1              1
#>  2              2
#>  3              3
#>  4              4
#>  5              5
#>  6              6
#>  7              7
#>  8              8
#>  9              9
#> 10             10
#> # … with 48 more rows
ir_sample_data[, c("id_sample", "spectra")]
#> # A tibble: 58 × 2
#>    id_sample spectra             
#>  * <chr>     <named list>        
#>  1 GN 11-389 <tibble [3,351 × 2]>
#>  2 GN 11-400 <tibble [3,351 × 2]>
#>  3 GN 11-407 <tibble [3,351 × 2]>
#>  4 GN 11-411 <tibble [3,351 × 2]>
#>  5 GN 11-416 <tibble [3,351 × 2]>
#>  6 GN 11-419 <tibble [3,351 × 2]>
#>  7 GN 11-422 <tibble [3,351 × 2]>
#>  8 GN 11-423 <tibble [3,351 × 2]>
#>  9 GN 11-428 <tibble [3,351 × 2]>
#> 10 GN 11-434 <tibble [3,351 × 2]>
#> # … with 48 more rows
ir_sample_data$id_sample
#>  [1] "GN 11-389"  "GN 11-400"  "GN 11-407"  "GN 11-411"  "GN 11-416" 
#>  [6] "GN 11-419"  "GN 11-422"  "GN 11-423"  "GN 11-428"  "GN 11-434" 
#> [11] "GN 11-435"  "GN 11-460"  "HW 07-151"  "HW 11-137"  "HW 11-144" 
#> [16] "HW 11-146"  "HW 11-148"  "HW 11-149"  "HW 11-172"  "HW 11-173" 
#> [21] "HW 11-176"  "HW 11-179"  "HW 11-184"  "LG 09-384"  "LG 11-309" 
#> [26] "LG 11-393"  "LG 11-395"  "LG 11-397"  "LG 11-404"  "LG 11-412" 
#> [31] "LG 11-414"  "LG 11-418"  "LG 11-432"  "LG 11-433"  "LG 11-438" 
#> [36] "LG 11-461"  "SW 11-138"  "SW 11-139"  "SW 11-154"  "SW 11-158" 
#> [41] "SW 11-178"  "OCC 11-457" "OCC 11-462" "OCC 11-463" "OMG 08-82" 
#> [46] "OMG 11-454" "OMG 11-455" "OMG 11-456" "ONP 08-78"  "ONP 09-388"
#> [51] "ONP 11-450" "ONP 11-451" "ONP 11-458" "ONP 11-459" "OFF 10-506"
#> [56] "OFF 13-144" "OFF 08-80"  "OFF 08-852"
#> attr(,"class")
#> [1] "character"
ir_sample_data[[1, 1]]
#> [1] 1
#> attr(,"class")
#> [1] "integer"

# setting and replacing columns
x <- ir::ir_sample_data
x$a <- 3
x[, "a"] <- 4
x$sample_type <- "a"
x[[1]] <- rev(x[[1]])

# deleting the spectra column drops the ir class
x$spectra <- NULL
class(x)
#> [1] "tbl_df"     "tbl"        "data.frame"

# setting and replacing rows
x <- ir::ir_sample_data
x[1, ] <- x[2, ]
class(x)
#> [1] "ir"         "tbl_df"     "tbl"        "data.frame"

# setting invalid values in the spectra column drops the ir class
x_replacement <- x[1, ]
x_replacement$spectra <- list(1)
x[1, ] <- x_replacement
class(x)
#> [1] "tbl_df"     "tbl"        "data.frame"

# setting and replacing values
x <- ir::ir_sample_data
x[[1, 1]] <- 100

# replacing an element in the spectra column by an invalid element drops the
# ir class attribute
x[[3, "spectra"]] <- list(1)
class(x)
#> [1] "tbl_df"     "tbl"        "data.frame"